fix(dashboard) rename dashboard file and add link styling
