+++ src/app/layout/DashboardLayoutBak.tsx
@@ -1,0 +1,223 @@
+import { useState } from 'react';
+import reduxSagaLogo from '../../assets/redux-saga.svg';
+import { NavLink, Outlet, useNavigate } from 'react-router-dom';
+import { ExitIcon } from '@radix-ui/react-icons';
+
+export default function DashboardLayout() {
+  const [isSideNavOpen, setIsSideNavOpen] = useState(false);
+  const navigate = useNavigate();
+  const handleSubmit = () => {
+    localStorage.clear();
+    navigate('/login');
+  };
+  return (
+    <>
+      {/*  <!-- Component: Side navigation menu with content separator --> */}
+      {/*  <!-- Mobile trigger --> */}
+      <button
+        title="Side navigation"
+        type="button"
+        className={`visible fixed left-6 top-6 z-40 order-10 block h-10 w-10 self-center rounded bg-white opacity-100 lg:hidden ${
+          isSideNavOpen
+            ? 'visible opacity-100 [&_span:nth-child(1)]:w-6 [&_span:nth-child(1)]:translate-y-0 [&_span:nth-child(1)]:rotate-45 [&_span:nth-child(3)]:w-0 [&_span:nth-child(2)]:-rotate-45 '
+            : ''
+        }`}
+        aria-haspopup="menu"
+        aria-label="Side navigation"
+        aria-expanded={isSideNavOpen ? true : false}
+        aria-controls="nav-menu-2"
+        onClick={() => setIsSideNavOpen(!isSideNavOpen)}
+      >
+        <div className="absolute top-1/2 left-1/2 w-6 -translate-x-1/2 -translate-y-1/2 transform">
+          <span
+            aria-hidden="true"
+            className="absolute block h-0.5 w-9/12 -translate-y-2 transform rounded-full bg-slate-700 transition-all duration-300"
+          ></span>
+          <span
+            aria-hidden="true"
+            className="absolute block h-0.5 w-6 transform rounded-full bg-slate-900 transition duration-300"
+          ></span>
+          <span
+            aria-hidden="true"
+            className="absolute block h-0.5 w-1/2 origin-top-left translate-y-2 transform rounded-full bg-slate-900 transition-all duration-300"
+          ></span>
+        </div>
+      </button>
+
+      {/*  <!-- Side Navigation --> */}
+      <aside
+        id="nav-menu-2"
+        aria-label="Side navigation"
+        className={`fixed top-0 bottom-0 left-0 z-40 flex w-72 flex-col border-r border-r-slate-200 bg-white transition-transform lg:translate-x-0 ${
+          isSideNavOpen ? 'translate-x-0' : ' -translate-x-full'
+        }`}
+      >
+        <a
+          aria-label="WindUI logo"
+          className="flex items-center gap-2 whitespace-nowrap p-6 text-xl font-medium focus:outline-none"
+          href="javascript:void(0)"
+        >
+          {/* <svg
+            width="300"
+            height="300"
+            viewBox="0 0 300 300"
+            fill="none"
+            xmlns="http://www.w3.org/2000/svg"
+            className="h-8 w-8 bg-emerald-500"
+          >
+            <path
+              fillRule="evenodd"
+              clipRule="evenodd"
+              d="M88.1121 88.1134L150.026 150.027L150.027 150.027L150.027 150.027L150.028 150.027L150.027 150.026L88.1133 88.1122L88.1121 88.1134ZM273.878 273.877C272.038 274.974 196.128 319.957 165.52 289.349L88.1124 211.942L26.1434 273.911C26.1434 273.911 -20.3337 196.504 10.651 165.519L88.1121 88.1134L26.1417 26.1433C26.1417 26.1433 69.6778 0.00338007 104.519 0H0V300H300V0H104.533C116.144 0.00112664 126.789 2.90631 134.534 10.651L211.941 88.1123L273.877 26.177C274.974 28.0159 319.957 103.926 289.349 134.535L211.942 211.942L273.878 273.877ZM273.878 273.877L273.912 273.857V273.911L273.878 273.877ZM273.877 26.177L273.911 26.1429H273.857C273.857 26.1429 273.863 26.1544 273.877 26.177Z"
+              fill="white"
+            />
+            <path
+              fillRule="evenodd"
+              clipRule="evenodd"
+              d="M0 0H300V300H0V0ZM150.026 150.025C121.715 99.731 88.1131 88.1122 88.1131 88.1122L10.6508 165.519C10.6508 165.519 26.143 150.027 150.026 150.027H150.027C150.026 150.027 150.026 150.027 150.026 150.027L150.026 150.027C99.731 178.339 88.1124 211.941 88.1124 211.941L165.52 289.348C165.52 289.348 150.032 273.86 150.027 150.027H150.029C178.341 200.323 211.944 211.942 211.944 211.942L289.352 134.535C289.352 134.535 273.864 150.023 150.027 150.027V150.027L150.027 150.027C200.322 121.715 211.941 88.1125 211.941 88.1125L134.534 10.651C134.534 10.651 150.026 26.1431 150.026 150.025ZM150.027 150.027L150.026 150.027C150.026 150.026 150.026 150.026 150.026 150.025C150.026 150.025 150.027 150.026 150.027 150.027ZM150.027 150.027L150.027 150.026L150.027 150.027C150.027 150.027 150.027 150.027 150.027 150.027L150.027 150.027ZM150.027 150.027C150.027 150.027 150.027 150.027 150.027 150.027H150.027L150.027 150.027Z"
+              fill="rgba(255,255,255,.2)"
+            />
+          </svg> */}
+          <img src={reduxSagaLogo} className="h-8" />
+          Saga Dashboard
+        </a>
+        <nav
+          aria-label="side navigation"
+          className="flex-1 divide-y divide-slate-100 overflow-auto"
+        >
+          <div>
+            <ul className="flex flex-1 flex-col gap-1 py-3">
+              <li className="px-3">
+                <NavLink
+                  to="/admin/dashboard"
+                  className="flex items-center gap-3 rounded p-3 text-slate-700 transition-colors hover:bg-emerald-50 hover:text-emerald-500 focus:bg-emerald-200  aria-[current=page]:bg-emerald-100 aria-[current=page]:text-emerald-600 "
+                >
+                  <div className="flex items-center self-center">
+                    <svg
+                      xmlns="http://www.w3.org/2000/svg"
+                      fill="none"
+                      viewBox="0 0 24 24"
+                      strokeWidth="1.5"
+                      stroke="currentColor"
+                      className="h-6 w-6"
+                      aria-label="Dashboard icon"
+                      role="graphics-symbol"
+                    >
+                      <path
+                        strokeLinecap="round"
+                        strokeLinejoin="round"
+                        d="M2.25 12l8.954-8.955c.44-.439 1.152-.439 1.591 0L21.75 12M4.5 9.75v10.125c0 .621.504 1.125 1.125 1.125H9.75v-4.875c0-.621.504-1.125 1.125-1.125h2.25c.621 0 1.125.504 1.125 1.125V21h4.125c.621 0 1.125-.504 1.125-1.125V9.75M8.25 21h8.25"
+                      />
+                    </svg>
+                  </div>
+                  <div className="flex w-full flex-1 flex-col items-start justify-center gap-0 overflow-hidden truncate text-sm">
+                    Dashboard
+                  </div>
+                </NavLink>
+              </li>
+              <li className="px-3">
+                <NavLink
+                  to="/admin/users"
+                  className="flex items-center gap-3 rounded p-3 text-slate-700 transition-colors hover:bg-emerald-50 hover:text-emerald-500 focus:bg-emerald-200 aria-[current=page]:bg-emerald-100 aria-[current=page]:text-emerald-600 "
+                >
+                  <div className="flex items-center self-center ">
+                    <svg
+                      xmlns="http://www.w3.org/2000/svg"
+                      fill="none"
+                      viewBox="0 0 24 24"
+                      strokeWidth="1.5"
+                      stroke="currentColor"
+                      className="h-6 w-6"
+                      aria-label="Dashboard icon"
+                      role="graphics-symbol"
+                    >
+                      <path
+                        strokeLinecap="round"
+                        strokeLinejoin="round"
+                        d="M15 19.128a9.38 9.38 0 002.625.372 9.337 9.337 0 004.121-.952 4.125 4.125 0 00-7.533-2.493M15 19.128v-.003c0-1.113-.285-2.16-.786-3.07M15 19.128v.106A12.318 12.318 0 018.624 21c-2.331 0-4.512-.645-6.374-1.766l-.001-.109a6.375 6.375 0 0111.964-3.07M12 6.375a3.375 3.375 0 11-6.75 0 3.375 3.375 0 016.75 0zm8.25 2.25a2.625 2.625 0 11-5.25 0 2.625 2.625 0 015.25 0z"
+                      />
+                    </svg>
+                  </div>
+                  <div className="flex w-full flex-1 flex-col items-start justify-center gap-0 overflow-hidden truncate text-sm">
+                    User Management
+                  </div>
+                </NavLink>
+              </li>
+            </ul>
+          </div>
+          <div>
+            <ul className="flex flex-1 flex-col gap-1 py-3">
+              <li className="px-3">
+                <NavLink
+                  to="/admin/reports"
+                  className="flex items-center gap-3 rounded p-3 text-slate-700 transition-colors hover:bg-emerald-50 hover:text-emerald-500 focus:bg-emerald-50 aria-[current=page]:bg-emerald-100 aria-[current=page]:text-emerald-600 "
+                >
+                  <div className="flex items-center self-center ">
+                    <svg
+                      xmlns="http://www.w3.org/2000/svg"
+                      fill="none"
+                      viewBox="0 0 24 24"
+                      strokeWidth="1.5"
+                      stroke="currentColor"
+                      className="h-6 w-6"
+                    >
+                      <path
+                        strokeLinecap="round"
+                        strokeLinejoin="round"
+                        d="M9 12h3.75M9 15h3.75M9 18h3.75m3 .75H18a2.25 2.25 0 002.25-2.25V6.108c0-1.135-.845-2.098-1.976-2.192a48.424 48.424 0 00-1.123-.08m-5.801 0c-.065.21-.1.433-.1.664 0 .414.336.75.75.75h4.5a.75.75 0 00.75-.75 2.25 2.25 0 00-.1-.664m-5.8 0A2.251 2.251 0 0113.5 2.25H15c1.012 0 1.867.668 2.15 1.586m-5.8 0c-.376.023-.75.05-1.124.08C9.095 4.01 8.25 4.973 8.25 6.108V8.25m0 0H4.875c-.621 0-1.125.504-1.125 1.125v11.25c0 .621.504 1.125 1.125 1.125h9.75c.621 0 1.125-.504 1.125-1.125V9.375c0-.621-.504-1.125-1.125-1.125H8.25zM6.75 12h.008v.008H6.75V12zm0 3h.008v.008H6.75V15zm0 3h.008v.008H6.75V18z"
+                      />
+                    </svg>
+                  </div>
+                  <div className="flex w-full flex-1 flex-col items-start justify-center gap-0 overflow-hidden truncate text-sm">
+                    Reports
+                  </div>
+                </NavLink>
+              </li>
+            </ul>
+          </div>
+        </nav>
+        <footer className="border-t border-slate-200 p-3">
+          <button
+            onClick={handleSubmit}
+            className="flex items-center gap-3 rounded p-3 text-slate-900 transition-colors hover:text-emerald-500 w-full"
+          >
+            {/* <div className="flex items-center self-center ">
+              <svg
+                xmlns="http://www.w3.org/2000/svg"
+                fill="none"
+                viewBox="0 0 24 24"
+                strokeWidth="1.5"
+                stroke="currentColor"
+                className="h-6 w-6"
+                aria-label="Dashboard icon"
+                role="graphics-symbol"
+              >
+                <path
+                  strokeLinecap="round"
+                  strokeLinejoin="round"
+                  d="M20.25 6.375c0 2.278-3.694 4.125-8.25 4.125S3.75 8.653 3.75 6.375m16.5 0c0-2.278-3.694-4.125-8.25-4.125S3.75 4.097 3.75 6.375m16.5 0v11.25c0 2.278-3.694 4.125-8.25 4.125s-8.25-1.847-8.25-4.125V6.375m16.5 0v3.75m-16.5-3.75v3.75m16.5 0v3.75C20.25 16.153 16.556 18 12 18s-8.25-1.847-8.25-4.125v-3.75m16.5 0c0 2.278-3.694 4.125-8.25 4.125s-8.25-1.847-8.25-4.125"
+                />
+              </svg>
+            </div> */}
+            <ExitIcon />
+            <div className="flex w-full flex-1 flex-col items-start justify-center gap-0 overflow-hidden truncate text-sm font-medium">
+              Logout
+            </div>
+          </button>
+        </footer>
+      </aside>
+      <div className="flex-1 ml-20 lg:ml-72 p-4">
+        <Outlet />
+      </div>
+
+      {/*  <!-- Backdrop --> */}
+      <div
+        className={`fixed top-0 bottom-0 left-0 right-0 z-30 bg-slate-900/20 transition-colors lg:hidden ${
+          isSideNavOpen ? 'block' : 'hidden'
+        }`}
+        onClick={() => setIsSideNavOpen(false)}
+      ></div>
+      {/*  <!-- End Side navigation menu with content separator --> */}
+    </>
+  );
+}
